Move slot processing into its own file
