+++ src/session.h
@@ -4,19 +4,11 @@
 #ifndef _SESSION_H
 #define _SESSION_H
 
-/* Slots */
-CK_RV p11prov_init_slots(P11PROV_CTX *ctx, P11PROV_SLOTS_CTX **slots);
-void p11prov_free_slots(P11PROV_SLOTS_CTX *slots);
-CK_RV p11prov_take_slots(P11PROV_CTX *ctx, P11PROV_SLOTS_CTX **slots);
-void p11prov_return_slots(P11PROV_SLOTS_CTX *slots);
-P11PROV_SLOT *p11prov_fetch_slot(P11PROV_SLOTS_CTX *sctx, int *idx);
-int p11prov_slot_get_mechanisms(P11PROV_SLOT *slot, CK_MECHANISM_TYPE **mechs);
-int p11prov_check_mechanism(P11PROV_CTX *ctx, CK_SLOT_ID id,
-                            CK_MECHANISM_TYPE mechtype);
-CK_RV p11prov_slot_get_obj_pool(P11PROV_CTX *provctx, CK_SLOT_ID id,
-                                P11PROV_OBJ_POOL **pool);
-
 /* Sessions */
+CK_RV p11prov_session_pool_init(P11PROV_CTX *ctx, CK_TOKEN_INFO *token,
+                                CK_SLOT_ID id, P11PROV_SESSION_POOL **_pool);
+void p11prov_session_pool_free(P11PROV_SESSION_POOL *pool);
+
 CK_SESSION_HANDLE p11prov_session_handle(P11PROV_SESSION *session);
 CK_SLOT_ID p11prov_session_slotid(P11PROV_SESSION *session);
 CK_RV p11prov_get_session(P11PROV_CTX *provctx, CK_SLOT_ID *slotid,
